Yeesh that was a pain: database full operational with services and formating doen
