+++ config/config.sh
@@ -49,38 +49,10 @@ yes_or_no() {
 display_asset() {
 	cat  $scriptPath/../assets/$1
 }
-print_boxes() {
-  local color_var=$1
-  shift
-  local texts=("$@")
-  local color=${!color_var}
-
-  # Print top borders
-  for text in "${texts[@]}"; do
-    local length=$( echo "$text" | sed -r 's/\\033\[[0-9;]*m//g' | wc -c )
-    #echo -n $( echo $text | sed -r 's/\\033\[[0-9;]*m//g')
-    echo -ne "${color}╔" && printf '═%.0s' $(seq 2 $length) && echo -n "╗ "
-  done
-  echo
-
-  # Print text values
-  for text in "${texts[@]}"; do
-    echo -ne "║${text}$color║ "
-  done
-  echo
-
-  # Print bottom borders
-  for text in "${texts[@]}"; do
-    local length=$( echo "$text" | sed -r 's/\\033\[[0-9;]*m//g' | wc -c )
-    echo -ne "╚" && printf '═%.0s' $(seq 2 $length) && echo -ne "╝ "
-  done
-  echo -e $RESTORE
-}
 
 export -f print_line
 export -f yes_or_no
 export -f display_asset
-export -f print_boxes
 
 
 #Has config run?
