+++ scripts/IPScrapper.sh
@@ -2,18 +2,8 @@
 
 # initialized by TY , modularized and refactored by D
 
-# Usage statement
-if [ -z "$1" ]; then
-	echo "Usage $0 <domain_name>"
-	exit 1
-fi
-
-# Config check
-if [ -z $CONFIG ]; then
-	echo "Command executed standalone: Running config"
-	export scriptPath=$(dirname $0)
-	source $scriptPath/../config/config.sh
-fi
+#Init code
+source $(dirname $0)/initTool.sh "Usage $0 <domain_name>" $@ 
 
 # Var definitions
 domain=$1
