+++ scripts/searchTargets.pl
@@ -5,6 +5,8 @@ use warnings;
 
 # Created by D
 
+#Init code (can't use one-liner because not in bash)
+
 # Usage statement
 if (@ARGV == 0) {
     print "Usage: $0 <search>\n";
@@ -41,14 +43,19 @@ else {
 		if ($target =~ /$search_term/)
 		{
 			# Extract the domain and IP
-my $ip = $target =~ /(\d{1-3}\.\d{1-3}\.\d{1-3}\.\d{1-3}:\d{1-5})/; #BUGGG idk perl regex syntax :(
-# Extract the ports
-my @ports = $target =~ /Port\s+(\d+):/g;
+			my @head = $target =~ /^.*\n/g;
+			# Extract the ports
+			my @ports = $target =~ /Port\s+(\d+):/g;
+			print "$target\n";
+			# Print the results
+			#print "head: @head\n";
+			#my $numPorts = scalar @targets;
+			#my $command = "$scriptPath/print_ports.py $numPorts \033[33m @ports";
+			#my $output = qx($command);
+			#print $output;
 
-# Print the results
-print "ip: $ip\n";
-print "Ports: @ports\n";
 		}
+		$is_blank = 1
 	}
 	if ($is_blank == 0) {
 		printf "\033[0;31mNo targets matched\033[0;37m '$search_term'\n"
